Fixed issue where HTTP to HTTPS redirect leaves out the /
Currently, when redirecting from HTTP to HTTPS, it fails if there is anything after $domain because it does not add back the `/`. 
Example: `http://example.com/apps/files/` redirects to `http://example.comapps/files/`
This is wrong!

I fixed it by adding the `/` on line 120.

Thanks for your time and consideration and have a great day!