Accept 90% of demanded RAM as sufficient
Usually not all RAM is available/listed in /proc/meminfo so a 2GB/8GB/whatever device usually has up to 10% less. The proposed change enables all devices labelled/marketed as a certain amount of GB to actually pass the test for this amount of GB. Otherwise none of my two tested devices pass the test for the amount of RAM that they are supposed to  have.

Signed-off-by: Lukas Bürgi <buergi.lukas@gmail.com>