Can elasticsearch container be bound only to localhost?
Is there any reason this container should expose ports to external network by default? 

I have seen that it's got the readonlyrest authenticating layer on it so I know it's not just "wide open", but it seems unnecessary from a security standpoint for it to be listening on 0.0.0.0 by default in this context?