Setting ramcheck back to 2
Originally set to 0.75 for RPi 3, but now back to testing on RPi4 4gb.  Although this may have to be commented out potentially.  Im not sure how much free Ram shows up on a 2gb model, I think it will only show about 1.8gb or so free so would still fail.  So we may have to lower this to 1gb if we want a ramcheck and just state that this will only work on RPi 2gb, 4gb and now new 8gb model.