Fixed reference to script name change
The name of the script change-to-zfs-mount-generator was changed to zfs-prune-snapshots.sh in commit f9c9e0d in /vm/disk. This breaks the install process as it fails to download. This PR changes the name to the correct name. Honestly I'm not too sure if this is the right way to do this but I couldn't find any other references to that script so hopefully it's correct.