allow the system to send mails
I though this can be useful.
Question is, if this should be enabled by default during the startup-script
and where exactly we can make use of this.
Also it could be in startup_configuration? Or does it fit good in here?
We could also move it to the top or so...

Otherwise I successfully tested this with a mailaccount.