switch to beta release channel if new major version was found
I think this makes sense since most of the time you won't see the new major update in https://your-nc-domain/settings/admin/overview unless you set it to the beta channel.
And as you know, this is no problem since the update script doesn't rely on this channel either way.
Signed-off-by: szaimen <szaimen@e.mail.de>