add btrfs options to the not-supported folder
This is the second of two PRs that allow BTRFS drives to be added into the backup scripts including scrubbing and snapshots as well as being the veracrypt encrypted.
This PR doesn't need as much review because it doesn't effect any official script. It may look like much but in I've basically copied from my other existing scripts.
Signed-off-by: szaimen <szaimen@e.mail.de>