Remove "Prefer IPv4 for apt"
I can understand the original intention, autonomous systems had issues with the current Internet Protocol a.k.a. IPv6 over the past years but this protocol is now very well established, mature and, IMO, there is no reason to a single application dictates how an operating system (or package manager in this regard) will behave regarding the system's Internet connectivity especially if the target environment has no issues with it and, if the opposite, masks any ongoing issue particular to the host OS, ISP and/or the server on the other end.

Circumventing things in the intention to get them done quick does not fix the root cause and it's better to put the effort once to fix things in a definitive way than sweeping it under the carpet to deal with it later and over and over. Any application shouldn't touch anything external to and/or broader than its own scope unless it's strictly necessary and the user is aware of any change & acknowledges it.

Signed-off-by: Anderson Silva <UnderEu@users.noreply.github.com>